+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changelog
 3.8.5 (2016-12-01)
 ------------------
 - Added support for CodeBuild by autodetecting a buildspec file and deploying with settings from that
+- Added support for working directories containing white-spaces
 
 ------------------
 3.8.4 (2016-11-16)