+++ CHANGES.rst
@@ -3,10 +3,15 @@ Changelog
 =========
 
 ------------------
+3.8.6 (2016-12-08)
+------------------
+- Added support for working directories containing white-spaces
+- Added ca-central-1 Canada (Central) support
+
+------------------
 3.8.5 (2016-12-01)
 ------------------
 - Added support for CodeBuild by autodetecting a buildspec file and deploying with settings from that
-- Added support for working directories containing white-spaces
 
 ------------------
 3.8.4 (2016-11-16)